.
实体中id的类型和basedao中不同，导致冲突，会造成上次的错误。已更正

+++ src/com/zm/dao/IBaseDao.java
@@ -12,7 +12,7 @@ public interface IBaseDao<T> {
 
 	public List<T> findall();
 
-	public T getById(int id);
+	public T getById(long id);
 
 	public List<T> getByIds(long[] ids);
 }
